refactor: move GLU operation implementation into separated file
